+++ NEWS.md
@@ -1,6 +1,9 @@
 # Change Log
 
 ## Unreleased
+### Changed
+- expand documentation of `transform_coordinates()` and `get_origin()`
+
 ### Deprecated
 - `geom_polar_ci()`, see vignette "polar-gam" for a working alternative